requirements.txt
#####ch2_openpyxl
et-xmlfile==1.1.0
jdcal==1.4.1
openpyxl==3.0.4

#####ch2_msoffcrypto-tool
cffi==1.14.6
cryptography==35.0.0
msoffcrypto-tool==4.12.0
olefile==0.46
pycparser==2.20

#####ch3_python-docx
lxml==4.6.3
python-docx==0.8.11

#####ch3_pywin32 (Windows Only)
#pywin32==303

#####ch4_bs4_html5lib
beautifulsoup4==4.9.1
html5lib==1.1
six==1.16.0
soupsieve==2.3.1
webencodings==0.5.1

#####ch4_requests
certifi==2021.10.8
chardet==3.0.4
idna==2.10
requests==2.24.0
urllib3==1.25.11

#####ch4_certifi
certifi==2021.10.8

#####ch4_ch6_flask
click==8.0.3
colorama==0.4.4
Flask==2.1.1
importlib-metadata==4.11.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
Werkzeug==2.0.2
zipp==3.8.0

#####ch4_chardet
chardet==3.0.4

#####ch4_selenium
selenium==3.141.0
urllib3==1.25.11

#####ch5_tweepy
certifi==2021.10.8
idna==2.10
oauthlib==3.1.1
PySocks==1.7.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.16.0
tweepy==3.9.0
urllib3==1.25.11

#####ch5_pyperclip
pyperclip==1.8.2

#####ch6_pysimplegui
PySimpleGUI==4.55.1

#####ch6_pyinstaller (Windows Only)
#altgraph==0.17.2
#future==0.18.2
#pefile==2021.9.3
#pyinstaller==4.7
#pyinstaller-hooks-contrib==2021.4
#pywin32-ctypes==0.2.0

#####ch6_pyautogui_opencv_pillow
MouseInfo==0.1.3
numpy==1.21.4
opencv-python==4.5.4.60
Pillow==8.4.0
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.28
pytweening==1.0.4
